+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ dev
     - Set ``SI – HU`` and ``SI – HU`` as Flow-based
 - Increase rate limits.
 - Increase default requests timeout.
+- Add ``proxies`` parameter in ``EnergyQuantified`` and ``RealtoConnection``
 
 **Bugfixes**